Fixed armor protection. Hopefully.

+++ common/net/silentchaos512/gems/api/tool/part/IPartProperties.java
@@ -27,7 +27,7 @@ public interface IPartProperties {
 
   public float getChargeSpeed();
 
-  public int getProtection();
+  public float getProtection();
 
   public ItemStack getCraftingStack();
